health-server: Adding get_default_value()
The yaml format allow people from defining global variables in the main
section or override them in each benchmark section.

The get_default_value() try to get the value of a parameter in a section
or use a default value.

This patch does implement this function and use it to get the default
runtime or required-hosts values.